Split output
