+++ src/ops/copy_object.rs
@@ -1,0 +1,60 @@
+//! [`CopyObject`](https://docs.aws.amazon.com/AmazonS3/latest/API/API_CopyObject.html)
+
+use super::*;
+use crate::dto::{CopyObjectError, CopyObjectOutput};
+
+impl S3Output for CopyObjectOutput {
+    fn try_into_response(self) -> S3Result<Response> {
+        wrap_output(|res| {
+            res.set_opt_header(|| X_AMZ_EXPIRATION.clone(), self.expiration)?;
+            res.set_opt_header(
+                || X_AMZ_COPY_SOURCE_VERSION_ID.clone(),
+                self.copy_source_version_id,
+            )?;
+            res.set_opt_header(|| X_AMZ_VERSION_ID.clone(), self.version_id)?;
+            res.set_opt_header(
+                || X_AMZ_SERVER_SIDE_ENCRYPTION.clone(),
+                self.server_side_encryption,
+            )?;
+            res.set_opt_header(
+                || X_AMZ_SERVER_SIDE_ENCRYPTION_CUSTOMER_ALGORITHM.clone(),
+                self.sse_customer_algorithm,
+            )?;
+            res.set_opt_header(
+                || X_AMZ_SERVER_SIDE_ENCRYPTION_CUSTOMER_KEY_MD5.clone(),
+                self.sse_customer_key_md5,
+            )?;
+            res.set_opt_header(
+                || X_AMZ_SERVER_SIDE_ENCRYPTION_AWS_KMS_KEY_ID.clone(),
+                self.ssekms_key_id,
+            )?;
+            res.set_opt_header(
+                || X_AMZ_SERVER_SIDE_ENCRYPTION_CONTEXT.clone(),
+                self.ssekms_encryption_context,
+            )?;
+            res.set_opt_header(|| X_AMZ_REQUEST_CHARGED.clone(), self.request_charged)?;
+
+            let copy_object_result = self.copy_object_result;
+
+            res.set_xml_body(64, |w| {
+                w.opt_stack("CopyObjectResult", copy_object_result, |w, result| {
+                    w.opt_element("ETag", result.e_tag)?;
+                    w.opt_element("LastModified", result.last_modified)
+                })
+            })?;
+
+            Ok(())
+        })
+    }
+}
+
+impl S3Output for CopyObjectError {
+    fn try_into_response(self) -> S3Result<Response> {
+        match self {
+            Self::ObjectNotInActiveTierError(msg) => {
+                XmlErrorResponse::from_code_msg(S3ErrorCode::ObjectNotInActiveTierError, msg)
+            }
+        }
+        .try_into_response()
+    }
+}
