+++ src/ops/put_object.rs
@@ -1,0 +1,42 @@
+//! [`PutObject`](https://docs.aws.amazon.com/AmazonS3/latest/API/API_PutObject.html)
+
+use super::*;
+use crate::dto::{PutObjectError, PutObjectOutput};
+
+impl S3Output for PutObjectOutput {
+    fn try_into_response(self) -> S3Result<Response> {
+        wrap_output(|res| {
+            res.set_opt_header(|| X_AMZ_EXPIRATION.clone(), self.expiration)?;
+            res.set_opt_header(|| ETAG, self.e_tag)?;
+            res.set_opt_header(
+                || X_AMZ_SERVER_SIDE_ENCRYPTION.clone(),
+                self.server_side_encryption,
+            )?;
+            res.set_opt_header(|| X_AMZ_VERSION_ID.clone(), self.version_id)?;
+            res.set_opt_header(
+                || X_AMZ_SERVER_SIDE_ENCRYPTION_CUSTOMER_ALGORITHM.clone(),
+                self.sse_customer_algorithm,
+            )?;
+            res.set_opt_header(
+                || X_AMZ_SERVER_SIDE_ENCRYPTION_CUSTOMER_KEY_MD5.clone(),
+                self.sse_customer_key_md5,
+            )?;
+            res.set_opt_header(
+                || X_AMZ_SERVER_SIDE_ENCRYPTION_AWS_KMS_KEY_ID.clone(),
+                self.ssekms_key_id,
+            )?;
+            res.set_opt_header(
+                || X_AMZ_SERVER_SIDE_ENCRYPTION_CONTEXT.clone(),
+                self.ssekms_encryption_context,
+            )?;
+            res.set_opt_header(|| X_AMZ_REQUEST_CHARGED.clone(), self.request_charged)?;
+            Ok(())
+        })
+    }
+}
+
+impl S3Output for PutObjectError {
+    fn try_into_response(self) -> S3Result<Response> {
+        match self {}
+    }
+}
