+++ src/utils/response.rs
@@ -1,12 +1,13 @@
 //! response util
 
-use crate::Response;
+use crate::{BoxStdError, Response};
 use hyper::{
     header::{self, HeaderName, HeaderValue, InvalidHeaderValue},
     Body, StatusCode,
 };
 use mime::Mime;
 use std::convert::TryFrom;
+use xml::{common::XmlVersion, writer::XmlEvent, EventWriter};
 
 /// `ResponseExt`
 pub trait ResponseExt {
@@ -25,6 +26,11 @@ pub trait ResponseExt {
 
     /// set status code
     fn set_status(&mut self, status: StatusCode);
+
+    /// set xml body
+    fn set_xml_body<F>(&mut self, cap: usize, f: F) -> Result<(), BoxStdError>
+    where
+        F: FnOnce(&mut EventWriter<&mut Vec<u8>>) -> Result<(), xml::writer::Error>;
 }
 
 impl ResponseExt for Response {
@@ -54,5 +60,27 @@ impl ResponseExt for Response {
 
     fn set_status(&mut self, status: StatusCode) {
         *self.status_mut() = status;
+    }
+
+    /// set xml body
+    fn set_xml_body<F>(&mut self, cap: usize, f: F) -> Result<(), BoxStdError>
+    where
+        F: FnOnce(&mut EventWriter<&mut Vec<u8>>) -> Result<(), xml::writer::Error>,
+    {
+        let mut body = Vec::with_capacity(cap);
+        {
+            let mut w = EventWriter::new(&mut body);
+            w.write(XmlEvent::StartDocument {
+                version: XmlVersion::Version10,
+                encoding: Some("UTF-8"),
+                standalone: None,
+            })?;
+
+            f(&mut w)?;
+        }
+
+        *self.body_mut() = Body::from(body);
+        self.set_mime(&mime::TEXT_XML)?;
+        Ok(())
     }
 }
